ml-projects/image-classifier/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.24.0
tensorflow>=2.13.0
opencv-python>=4.8.0
pillow>=9.5.0 